refactor(cli): update.rs を mod.rs + ops.rs + tests.rs に分割 (SPEC-1942 SC-027)
SPEC-1942 SC-027 (200-600 行) を update family で完全達成。 update.rs (786
行) を 3 sub-file に分割:

- cli/update/mod.rs (250 行): UpdateRunMode + RunOutcome + parse_args +
  run_with + run + run_internal_apply_update + run_internal_run_installer +
  parse_flag_* helper。 SC-027 範囲内。
- cli/update/ops.rs (131 行): UpdateCliOps trait + RealUpdateCliOps struct +
  Default impl + UpdateCliOps impl。 trait と production 配線を分離。
- cli/update/tests.rs (413 行、 cfg(test)): FakeUpdateCliOps mock + 全
  test fn。 #[cfg(test)] mod tests; 経由で mod.rs から取り込み。

trait + RealUpdateCliOps は pub(super) として ops.rs から mod.rs に再公開。
mod.rs に ops 直接依存 (run_with の trait bound、 run() の RealUpdateCliOps
インスタンス化) のみ通す。 既存 caller (cli.rs から呼ぶ
update::UpdateRunMode / run / run_internal_*) は不変。

機能・契約・test 結果は不変 (lib 345 全 pass、 clippy 0 / fmt clean)。
